requirements.txt
Django==2.1.1
django-filter==2.0.0
pytz==2018.5
